Tarea #633 - exportAction imprime solo pestaña activa
